fix: integrate aqi and weather
